:recycle: Refactor: SignupSection2.jsx- 함수 @hoooks로 빼고 파일 경로 alias로 변경
