+++ src/components/signup/SignupSection2.jsx
@@ -1,10 +1,11 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styles from './SignupSection.module.css';
-import { handleSignup } from '../../utils/register.js';
-import { handleInputChange } from '../../utils/inputOnChange.js';
-import { APIService } from '@api/axios';
+import { handleInputChange } from '@utils/inputOnChange.js';
 import ConsentTable from './ConsentTable';
+
+import { handleSelectBox, handleBlurSelcetBox, handlePart } from '@hooks/useSignupDropdownHook.js';
+import { handleCheckboxChange, handleSignupClick } from '@hooks/useSignupHook.js';
 
 export default function SignupSection({ email, setSignupSuccess, setNow }) {
   const fullEmail = `${email}@skuniv.ac.kr`;
@@ -26,83 +27,6 @@ export default function SignupSection({ email, setSignupSuccess, setNow }) {
   const [isDropdownView, setIsDropdownView] = useState(false);
   const [selcetPart, setSelectPart] = useState('파트 선택');
   const navigate = useNavigate();
-
-  // email이 변경될 때마다 form의 id 업데이트
-  useEffect(() => {
-    setForm((prev) => ({ ...prev, id: `${email}@skuniv.ac.kr` }));
-  }, [email]);
-
-  function handleCheckboxChange(event) {
-    setForm({ ...form, consent: event.target.checked });
-  }
-
-  // 파트(part) 드롭아웃 메뉴 관련 함수
-  function handleSelectBox(event) {
-    event.preventDefault;
-    setIsDropdownView(!isDropdownView);
-  }
-
-  function handleBlurSelcetBox() {
-    setTimeout(() => {
-      setIsDropdownView(false);
-    }, 100);
-  }
-
-  function handlePart(event) {
-    if (form.part === event.target.id) {
-      setForm({ ...form, part: '' });
-      setSelectPart('파트 선택');
-    } else {
-      setForm({ ...form, part: event.target.id });
-      if (event.target.id === 'PM/design') {
-        setSelectPart('기획/디자인');
-      } else if (event.target.id === 'front') {
-        setSelectPart('프론트앤드');
-      } else if (event.target.id === 'back') {
-        setSelectPart('백앤드');
-      }
-    }
-  }
-
-  // 회원가입 버튼 클릭 //
-  function handleSignupClick(event) {
-    event.preventDefault();
-
-    const isValid = handleSignup(setErrors, form);
-    if (isValid === true && form.id_valid === true && form.consent === true) {
-      signUp();
-    }
-  }
-
-  async function signUp() {
-    try {
-      const requestData = {
-        loginId: form.id,
-        password: form.password,
-        userName: form.name,
-        department: form.department,
-        studentId: form.strudent_num,
-        semester: form.semester === '' ? 0 : Number(form.semester),
-        phoneNumber: form.phone_num,
-        parts: form.part,
-      };
-
-      const response = await APIService.public.post(import.meta.env.VITE_APP_SIGN_UP, requestData);
-
-      if (response.success) {
-        setSignupSuccess(true);
-        setNow(1);
-        navigate(`/welcome?name=${encodeURIComponent(form.name)}`);
-      } else {
-        let tologin = confirm(response.message + ' 로그인 페이지로 이동합니다.');
-        if (tologin) {
-          navigate('/login');
-        }
-      }
-    } catch {
-      alert('회원가입 중 서버 오류가 발생했습니다. 나중에 다시 시도해주세요');
-    }
-  }
 
   return (
     <div className={styles['signup-form']}>
@@ -281,10 +205,14 @@ export default function SignupSection({ email, setSignupSuccess, setNow }) {
           </div>
           <div
             className={styles['signup-form__inputsection']}
-            onBlur={handleBlurSelcetBox}
+            onBlur={handleBlurSelcetBox(setIsDropdownView)}
           >
             <div className={styles['signup-form__selectsection']}>
-              <label onClick={handleSelectBox}>
+              <label
+                onClick={function (event) {
+                  handleSelectBox(event, isDropdownView, setIsDropdownView);
+                }}
+              >
                 <button
                   id='part'
                   style={{ cursor: 'pointer' }}
@@ -308,7 +236,9 @@ export default function SignupSection({ email, setSignupSuccess, setNow }) {
                       : styles['signup-form__selection']
                   }
                   id='PM/design'
-                  onClick={handlePart}
+                  onClick={function (event) {
+                    handlePart(event, form, setForm, setSelectPart);
+                  }}
                 >
                   기획/디자인
                 </li>
@@ -319,7 +249,9 @@ export default function SignupSection({ email, setSignupSuccess, setNow }) {
                       ? styles['signup-form__selectionSelected']
                       : styles['signup-form__selection']
                   }
-                  onClick={handlePart}
+                  onClick={function (event) {
+                    handlePart(event, form, setForm, setSelectPart);
+                  }}
                 >
                   프론트앤드
                 </li>
@@ -330,7 +262,9 @@ export default function SignupSection({ email, setSignupSuccess, setNow }) {
                       ? styles['signup-form__selectionSelected']
                       : styles['signup-form__selection']
                   }
-                  onClick={handlePart}
+                  onClick={function (event) {
+                    handlePart(event, form, setForm, setSelectPart);
+                  }}
                 >
                   백앤드
                 </li>
@@ -347,7 +281,9 @@ export default function SignupSection({ email, setSignupSuccess, setNow }) {
               id='consent'
               checked={form.consent}
               className={errors.consent ? styles['invalid'] : form.consent ? styles['valid'] : ''}
-              onChange={handleCheckboxChange}
+              onChange={function (event) {
+                handleCheckboxChange(event, form, setForm);
+              }}
             ></input>
             <label htmlFor='consent'>(필수) 개인정보 수집 및 이용 동의서</label>
             <p>*</p>
@@ -361,7 +297,9 @@ export default function SignupSection({ email, setSignupSuccess, setNow }) {
           <button
             style={{ cursor: 'pointer' }}
             className={styles['signup-form__button--submitting']}
-            onClick={handleSignupClick}
+            onClick={function (event) {
+              handleSignupClick(event, form, setErrors, form, setSignupSuccess, setNow, navigate);
+            }}
           >
             회원가입
           </button>
